Creating MistakeSigns component
